src/requirements.txt
colorama
crossfiledialog
ffmpeg
pillow